move mapper into internal/index
